refactor: make fold4d a ModuleOps default method dispatched per-backend
